refactor: initial migration of the server code
