--- packages/core/server/content-api/controllers/url-alias.ts
@@ -1,10 +1,0 @@
-
-import { factories } from '@strapi/strapi';
-
-/**
- * URL alias controller
- */
-
-const contentTypeSlug = 'plugin::webtools.url-alias';
-
-export default factories.createCoreController(contentTypeSlug);
