--- packages/core/server/content-api/routes/core.ts
@@ -1,12 +1,0 @@
-
-
-export default [
-  {
-    method: 'GET',
-    path: '/router',
-    handler: 'core.router',
-    config: {
-      policies: [],
-    },
-  },
-];
