--- packages/core/server/content-api/routes/url-alias.ts
@@ -1,12 +1,0 @@
-
-
-export default [
-  {
-    method: 'GET',
-    path: '/url-alias',
-    handler: 'url-alias.find',
-    config: {
-      policies: [],
-    },
-  },
-];
