--- packages/core/server/content-api/services/url-alias.ts
@@ -1,9 +1,0 @@
-import { factories } from '@strapi/strapi';
-
-/**
- * URL alias service
- */
-
-const contentTypeSlug = 'plugin::webtools.url-alias';
-
-export default factories.createCoreService(contentTypeSlug);
