+++ packages/core/server/content-types/url-alias/schema.json
