+++ packages/core/server/content-types/url-pattern/schema.json
