+++ packages/core/server/controllers/index.ts
@@ -1,0 +1,11 @@
+import urlAliasController from './url-alias';
+import urlPatternController from './url-pattern';
+import infoController from './info';
+import coreController from './core';
+
+export default {
+  'url-alias': urlAliasController,
+  'url-pattern': urlPatternController,
+  info: infoController,
+  core: coreController,
+};
