+++ packages/core/server/register.ts
@@ -2,17 +2,13 @@
 /* eslint-disable @typescript-eslint/no-unsafe-member-access */
 
 import set from 'lodash/set';
-import { Schema } from '@strapi/strapi';
-import { IStrapi } from '../types/strapi';
-import { isContentTypeEnabled } from '../util/enabledContentTypes';
-import migratePluginOptionsRename from './migrations/plugin-options-rename';
+import { Core, Schema } from '@strapi/strapi';
+import { IStrapi } from './types/strapi';
+import { isContentTypeEnabled } from './util/enabledContentTypes';
 import { disableContentType } from './hooks/disable';
 
-export default (strapi: IStrapi) => {
-  // Migrate the pluginOptions to reflect the plugin rename.
-  migratePluginOptionsRename(strapi);
-
-  strapi.hook('strapi::content-types.beforeSync').register(disableContentType);
+export default ({ strapi }: { strapi: Core.Strapi }) => {
+  // strapi.hook('strapi::content-types.beforeSync').register(disableContentType);
 
   // Register the url_alias field.
   Object.values(strapi.contentTypes).forEach((contentType: Schema.ContentType) => {
