+++ packages/core/server/services/index.ts
@@ -1,0 +1,11 @@
+import urlAliasController from './url-alias';
+import urlPatternController from './url-pattern';
+import bulkGenerateController from './bulk-generate';
+import queryLayerDecorator from './query-layer-decorator';
+
+export default {
+  'url-alias': urlAliasController,
+  'url-pattern': urlPatternController,
+  'bulk-generate': bulkGenerateController,
+  'query-layer-decorator': queryLayerDecorator,
+};
